// src/app/models/User.ts
export interface User {
  uid: string;
  displayName: string;
  email: string;
  phoneNumber: string;
  photoURL: string;
  providerId: string;
  company:string;
  job_title:string;
  city:string;
  address:string;
  country:string;
  about:string;
  department:string;
  emailVerified:boolean;
  contractor: boolean;
}